fpm-compatibility initial commit
The project was restuctured to enable the use of fpm as a build system
and dependency manager for DNAOD.

NOTE: The current version has only been tested with gfortran. To enable
parallel compatibility with ifx, the used of cpp for preprocessing could
be explored in the future. CPP support may be enabled within fpm by
adding the following lines to the `fpm.toml` file:

```toml
[preprocess]
[preprocess.cpp]
```